// src/interfaces/Circunferencia.ts
export interface Circunferencia {
    id: string | number;
    data?: string;
    altura?: number;
    peso?: number;
    imc?: number;
    ombro?: number;
    cintura?: number;
    quadril?: number;
    peitoral?: number;
    abdomen?: number;
    coxaProximalEsquerda?: number;
    coxaProximalDireita?: number;
    coxaMedialEsquerda?: number;
    coxaMedialDireita?: number;
    coxaDistalEsquerda?: number;
    coxaDistalDireita?: number;
    panturrilhaEsquerda?: number;
    panturrilhaDireita?: number;
    bracoRelaxadoEsquerdo?: number;
    bracoRelaxadoDireito?: number;
    bracoContraidoEsquerdo?: number;
    bracoContraidoDireito?: number;
    antebraçoEsquerdo?: number;
    antebraçoDireito?: number;
}